Prettifying
Fixed extra white space as the bottom of the content pages. Fixed the
borders on the rotator pages. Changed the Welcome message to South Hills
is excited to be a part of InfoComm 2016, right here in Las Vegas! Fixed
the © issue at the bottom of the rotator pages. Added images to low
content pages.

-Jim